Fix  pbm list layout order
Fix for PBM-438 changed the listing order so the most recent backups became at the top of the list. This commit restores the listing order returning backup the behaviour when the most recent backups are listed in the bottom.

was
$ pbm list
Backup history:
  2020-04-03T12:38:26Z
  2020-03-17T20:37:39Z
  2020-03-17T19:37:40Z

become:
 pbm list
Backup history:
  2020-03-17T19:37:40Z
  2020-03-17T20:37:39Z
  2020-04-03T12:38:26Z

It also fixes the restore listing.