PBM-520 fix: backup's meta should be unique
If some node is lagging after it recived backup command it may "wake up" after the backup if fact was already done. By the node we try make a new one. Whicn in turn will create a mess.

1. Backup's meta collection now have a uinique index by backup name. So trying to create "a new" meta will result in error and this redundant attempt will fail. But this prevents only leader node from starting the backup

PBM-520 fix: backup's meta should be unique

If some node is lagging after it received backup command it may "wake up" after the backup, in fact, was already done. But the node will try to make a new one. Which in turn will create a mess.

1. Backup's meta collection now has a unique index by backup name. So trying to create "a new" meta will result in an error and this redundant attempt will fail. But this prevents only the leader node from starting the backup

2. We also should prevent lagging nodes from spoiling a valid backup meta with the errors of a failed attempt. For that, some extra checks have been added to the change meta status procedure.