PBM-485 fix: Upload 100gb+ db on aws
AWS MaxUploadParts is 1e4 so with PartSize = 10Mb the max allowed file size
would be ~ 97.6Gb. So if the file size is bigger we're enlarging PartSize
so PartSize * MaxUploadParts could fit the file.
If calculated PartSize is smaller than the default we leave the default.
sizeb usually would be an uncompressed db size on disk.

In fact, for all DBs the upload size would be the same
and equals to the size of the largest DB. We're ok with that
since there are no easy way to define (distinct) DBs in upload
func given we're piping dump thru the mongodump util.