PBM-485 fix: Upload 100gb+ db on aws (#417)
MaxUploadParts is 1e4 so with PartSize 10Mb the max allowed
file size would be ~ 97.6Gb. Hence if the file size is bigger we're
enlarging PartSize so PartSize * MaxUploadParts could fit the file. 
If calculated PartSize is smaller than the default we leave the default.
If UploadPartSize option was set we use it instead of the default. 
Even with the UploadPartSize set the calculated PartSize woulbe used if it's bigger.

If backup won't be compressed we're assuming that the dump size
could be up to 4x larger due to mongo's wieredtiger compression.
Otherwise, we assume that it would be covered by our compression. 
But in any case, there is an extra 10% gap while calculating chunk size.