PBM-537 feature: check oplog sufficient range
Dusring the oplog slicing, before processing the first
oplog record we check if oplog has sufficient range,
i.e. if there are no gaps between the ts of the last backup or slice and
the first record of the current slice. Whereas the request is >= last_saved_ts
we don't know if the returned oldest record is the first since last_saved_ts or
there were other records that are now removed because of oplog collection capacity.
So after we retrieved the first record of the current slice we check if there is
at least one preceding record (basically if there is still record(s) from the previous set).
If so, we can be sure we have a contiguous history with respect to the last_saved_slice.

We should do this check only after we retrieved the first record of the set. Otherwise,
there is a possibility some records would be erased in a time span between check and first
recorded retrieval due to ongoing write traffic (i.e. oplog append). There's a chance of
false-negative though.

https://jira.percona.com/browse/PBM-537