PBM-460: Prevent conflicting namespace during the restore
Restore should remove tmp user and role collections otherwise
it would be a conflicting namespaces error during restore

This commit addresses issue by
1. Ensuring that restore would try to drop tmp collections even on error.
Which didn't happened before. The drop was run only on success.

2. Backup would ensure there ate no tmp collections and try to drop it otherwise.

There is no way to exclude some collections on mongodump stage
while it accepts "exclude..." option only if sole db specified (¯\_(ツ)_/¯)
So we're trying to clean up on primary node when it gets bcp command
but because there is no guarantee
that primary node reaches that code before some of the secondaries starts backup
we have to double-check in backup.run() before mongodump and fail on the dirty state.