PBM-595 fix: sharded backup with different rs name
`_id` in `config.shards` collection may differ from rs name.
So the `pbm-agent` would register itself in the backup meta by the rs name.
While peers would search it by the `_id` field of `config.shards`.
This commit changes the rs name in shards info to be the rs name,
not the shard id.