PBM-615: check if backups match the cluster
During `pbm list` and `pbm status` we compare each backup with the
current cluster and if backup's data doesn't match cluster backup's meta
would be changed to pbm.StatusError with respective error text emitted. It doesn't change meta on
storage nor in DB (backup is ok, it just doesn't cluster), it is just "in-flight" changes
in given `bcps`.

Match means that for every shard in the cluster there is data in the backup.
It's ok if the backup has data for more shards than there are currently in cluster.

https://jira.percona.com/browse/PBM-615