PBM-615: check if backups match the cluster (#490)
This commit changes the restore behaviour regarding backup - target cluster match

Now match means that each replset in the backup has respective replset on the target cluster. It's ok if the
cluster has more shards than there are currently in backup.

Also during pbm list and pbm status we compare each backup with the
current cluster and if backup's data doesn't match cluster backup's meta
would be changed to pbm.StatusError with respective error text emitted. It doesn't change meta on
storage nor in DB (backup is ok, it just doesn't cluster), it is just "in-flight" changes
in given bcps.

https://jira.percona.com/browse/PBM-615

Co-authored-by: Akira Kurogane <akira.kurogane@gmail.com>