PBM-636: fix waiting time for tmp users
We were writing to the primary but reading last_write of the majority.
So time_to_wait usually were some previous write because the needed one
wasn't replicated yet to the majority.

Last_write to wait for should be taken not from the majority but from
the node we are writing tmp users to.