PBM-543 feature: Configurable duration PITR slices' span (#553)
Set in minutes via config option `pitr.oplogSpanMin`. Option is float so fractures of minutes
could be set. E.g. `0.5` is 30 sec.

PITR routine every 15 seconds along with `pitr.enabled` would be check `pitr.oplogSpanMin`.
And if the agent runs slicer and span duration has changed, the slicer would be
updated with the new value. If the new value is smaller, the slicer also
would be woken up to immediately make a slice and reset the timer.
If the new value is bigger, then the slicer won't be woken up (it will apply
a new value only after the current span finishes)