PBM-617 PITR: ability to restore from previous snapshots (#555)
Backups no more split the PITR timeline. So point-in-time recovery can be based
on any preceding backup, not only the most recent in respect to the desired time.

Prior to this commit, given
F1 a b c d e f F2 g h i j k l m F3 n
broken F2 would compromise timeline g-m. But now say time k can be recovered
based on backup F1.

To achieve this, restarting after the backup the PITR catchup process would
copy the snapshot's oplog as the PITR chunk (using storage copy, not 
download/uploading it). Also, the catchup will try to fill the gap between
the last chunk (complete time of the oldest backup if there is no chunk) and
the new starting point. It will do so by moving backwards in time and copying
snapshots' oplogs and making PITR-slices in between backups. The catchup would
stop either reaching the last chunk or if find the oplog insufficiency of
restore. So the PITR-timeline can be split only by a restore or the oplog
insufficiency (say restarting PITR after a while). If insufficiency or restore
found the catchup shall not do any slices up to the previous (in time) backup
as well as no oplog copies of previous backup. Since the previous backup in
such case is the oldest stable base and nothing prior to its complete time
can't belong to the following timeline.

To pbm restore added flag --base-snapshot in case of using --time (pitr restore).
Without --base-snapshot PBM would take the most recent snapshot in respect to
the given --time.
Example: pbm restore --time="2021-06-10T17:05:23" --base-snapshot="2021-06-09T17:24:49Z"

pbm delete-backup now deletes only snapshots not touching any PITR chunks.
Though it won't delete snapshot(s) that is base for any valid timeline. In the
example above F1 won't be deleted, but F2 and F3 will. Also, it won't delete
the most recent backup if PITR is ON and there are no chunks after the snapshot
yet. In this case, we yet don't know if the backup gonna be the base of the
following timeline.

Deleting of the PITR slices will be added in https://jira.percona.com/browse/PBM-652

Adjusting CI-tests in https://jira.percona.com/browse/PBM-688

https://jira.percona.com/browse/PBM-617

Co-authored-by: Akira Kurogane <akira.kurogane@gmail.com>

Co-authored-by: Akira Kurogane <akira.kurogane@gmail.com>