PBM-617: fix timelines merge and status layout
PITR layout in status now shows the total size of all chunks. Previous
the layout might be misleading as we show not files but available time
ranges (conjunction of all sets) so some ranges which happened to be
out of final timelines are still present on storage.

Before:
```
...
  PITR chunks [3.33MB]:
    2021-06-28T08:19:55 - 2021-06-28T09:11:06
    2021-06-23T15:40:42 - 2021-06-23T16:38:16
```

Now:
```
...
  PITR chunks:
    2021-06-28T08:19:55 - 2021-06-28T09:11:06  1.80MB
    2021-06-23T15:40:42 - 2021-06-23T16:38:16  1.53MB
```

Also PBM-688: fix delete tests