PBM-652: delete PITR chunks (#565)
* PBM-652: delete PITR chunks

Added command `pbm delete-pitr`. With two flags:
- ` -a, --all` - delete all chunks
- `--older-than`

If `--older-than` is set it will be rounded down to the last write of the closest preceding backup in order not to make gaps. So usually it
gonna leave some extra chunks. E.g.older-than = 13 and the last write of the closest preceding
backup is 10 it will leave 11 and 12 chunks as well since 13 won't be restorable without 11 and
12 (contiguous timeline from the backup).


https://jira.percona.com/browse/PBM-652