PBM-700 fix: pitr/backup oplogs aligment
It can happen that by the time of the backup/restore sync there is already a chunk with the end time bigger than the backup first_write.
To make it worse RS's first_write could be much later than the backups one (backup's - is the earliest of all RS's).
Then Slicer creates the last pre-backup slice which start_time is bigger than the backups first_write and the end-time smaller than the backups last_write.
Finally, resuming slicing after the backup catchup process will copy the backup's oplog as the pitr chunk.
So in the end there are would be one chunk that covers a subset of the copied from the backup one. But that former chunk is perfectly aligned with the end of the breakup timeline. So PBM will discard the backup's chunk and we got timeline split when it shouldn't be.
example:
```
I [pitr] created chunk 2021-08-06T08:17:32 - 2021-08-06T08:18:32

rs first_rwite: 2021-08-06T08:18:39
backup first_write: 2021-08-06T08:18:28

I [pitr] created chunk 2021-08-06T08:18:32 - 2021-08-06T08:18:39

... backup ...

I [pitr] copied chunk 2021-08-06T08:18:28 - 2021-08-06T08:18:41

have next time frames for the RS:
{ 2021-08-06T08:03:09 - 2021-08-06T08:18:39} { 2021-08-06T08:18:41 - 2021-08-06T08:33:17}
```

To avoid this:
1. The last pre-backup chunks should ends on the backups fisrt_write, not the rs' fisrt_write.
2. Making the last chunk slicer would check if backup's last_write is bigger than the end-time of the last saved chunks. Otherwise, it shouldn't do anything, just exit.