PBM-705: don't write an error when the reader close the pipe (#582)
The last chunk during the PITR restore usually won't be read fully
(high chances that the targeted time will be in the middle of the chunk)
so in this case the reader (oplog.Apply) will close the pipe once reaching the
targeted time. And this is no case for logging the error.