PBM-725: oplog: don't replay config.rangeDeletions
Before starting the backup PBM disables a balancer (if it was on). So it
shouldn’t be any rangeDeletions ops in the oplog. But they are there because
oplog can capture some events from before the mongodump started. Simply because
we don’t know the cluster time of the mongodump start so replicasets do just
rely on the oldest “last write” among them as the first write for the oplog to
capture. So given that we are turning off the balancer before mongodump, all
`config.rangeDeletions` records in oplog have happened before hence they are
in the mongodump output. So basically it’s safe to skip such records during the
oplog replay.

https://jira.percona.com/browse/PBM-725