PBM-735: fix race in PITR catchup (#606)
It may happen that due to the small and fast backup the pre-backup oplog slicer
gets into a race condition with the post-backup one. Both are trying to save
an oplog chunk that ends at backup_start_time. If post-backup fails to do so
(the pre-backup one had already done that) it will fail with
the `E11000 duplicate key error collection: admin.pbmPITRChunks` error and
won't make the PITR slice that covers backup time. Which will result in a gap
in the oplog timeframe.

To fix this catchup process (post-backup slicer) should not fail due to
`duplicate key error`

Fix: https://jira.percona.com/browse/PBM-735