PBM-805: Physical backups (#638)
Physical backup functionality is implemented using the Backup cursor https://www.percona.com/blog/2021/06/07/experimental-feature-backupcursorextend-in-percona-server-for-mongodb/. Algorithm of the backup is pretty much following the article.

For the restore, unlike in logical restore, _every_ node of the replicaset is taking part in
a physical restore. In that way, we avoid logical resync between nodes after
the restore. Each node in the cluster does:
- Stores current replica set config and mongod port.
- Checks backup and stops all routine writes to the db.
- Stops mongod and wipes out datadir.
- Copies backup data to the datadir.
- Starts standalone mongod on ephemeral (tmp) port and reset some internal
  data also setting one-node replicaset and sets oplogTruncateAfterPoint
  to the backup's `last write`. `oplogTruncateAfterPoint` set the time up
  to which journals would be replayed.
- Starts standalone mongod to recover oplog from journals.
- Cleans up data and resets replica set config to the working state.
- Shuts down mongod and agent (the leader also dumps metadata to the storage).