PBM-805: trun off pitr after the physical restore
PITR should be turned off after the physical restore. Otherwise, if it was on during
the backup, after the user runs the cluster again the slicing resumes in the state
of the backup's time. Hence the system knows nothing about the recent restore and chunks
made after the backup. So it would successfully start slicing and would overwrite chunks
after the backup.