PBM-809: Restore pitr chunks - try s2 if snappy failed (#650)
Up until v1.7.0 the compression of pitr chunks was always S2.
But it was a mess in the code which lead to saving pitr chunk files
with the `.snappy` extension although it was S2 in fact. And during
the restore, decompression treated .snappy as S2 ¯\_(ツ)_/¯ It wasn’t
an issue since there was no choice. Now, Snappy produces `.snappy` files
and S2 - `.s2` which is ok. But this means the old chunks (made by previous
PBM versions) won’t be compatible - during the restore, PBM will treat such
files as Snappy (judging by its suffix) but in fact, they are s2 files,
and restore will fail with snappy: corrupt input. So we try S2 in such a case.