PBM-896: add status "partlyDone" for physical restore
Since we restore physical bcps on all nodes in the replset, we may treat
a success on at least one node in each replset as a partial success.
Meaning the cluster can be started with the restore although some nodes
gonna be in initiallSync.

For that, if at least one node in the replset moved
to the "done", the replset is "partlyDone". And a replset is "done" if all
nodes moved to "done". For cluster success, all replsets must move either
to "done" or "partlyDone". Cluster
is "partlyDone" if at least one replset
is "partlyDone".

Also `pbm describe-restore` shows the restore status accordingly.