PBM-978: fix phys restore of hidden nodes
When bson value is "omitempty" it won't passi to the mongo and mongo
writes default. As the default for float64 is 0, it will be omited by
driver (omitempty) and mongo will write its default for an option.
And default for `priority` is 1. So when it was 0 it became 1 after
the PBM write.
This commit fixes it.

https://jira.percona.com/browse/PBM-978