PBM-999: Add Incremental Backpus (#740)
Based on $backupCursor and WT block incremental backups features

First of all, the user would need to make a full incremental backup (a base).
The node will start keeping an incremental backup history. And when a
subsequent incremental buck will be requested, the system will be able to
calculate and provide the difference in data blocks after the last incr backup.
So PBM would store only the diffs for every next incr backup. For the restore,
PBM would basically retrieve all the preceding backups chain up to the closest
base, do a physical restore of the base backup and then advance forward in time
(from the oldest after the base incr backup in the chain) and apply diffs onto
restore files.

CLI interface updates:
```
$ pbm backup -type incremental --base // do an incremental base backup 
$ pbm backup -type incremental // do an incremental backup based on the previous one
```

WIP: this PR bring only incremental backups and restores. If a subsequent
incremental backup goes to a different node, it will fail. Also need to deal
with deletes (deletion of any src will invalidate succeeding backups) and
testing etc.

https://jira.percona.com/browse/PBM-999