PBM-1047: fix physical restore to new sharded cluster (#789)
A `shardIdentity` document in `admin.system.version` collection on
shards contains the config server connection string.
A `config.shards` on the config server contains information about
shards. This info should be updated during the restore with data from
the target cluster. Otherwise, shards won't start after the restore if
the target is a cluster with new (compared to the backup) IPs/hostnames