PBM-1058: fix affected backups during restore (#807)
Checks if dbpath exists in the file name and cut it from the destination.
We suppose that "journal" will always be present in the backup and it is
always in the dbpath root and doesn't contain subdirs, only files. Having
a leading `/` indicates that restore was affected by PBM-1058 but only
with journal files we may detect the exact prefix.