PBM-1063: handle directoryPerDB for incremental backups  (#811)
Incremental $backupCusor returns collections that were created but
haven't ended up in the checkpoint yet in the format if they don't
belong to this backup (see PBM-1063). PBM doesn't copy such files. But
they are already in WT metadata. PSMDB (WT) handles this by creating
such files during the start. But fails to do so if it runs with
`directoryPerDB` option. Namely fails to create a directory for the
collections. So we have to detect and create these directories during
the restore.