PBM-1426: Sporadic failures on physical restore with PITR (#1061)
Ensure that oplog is applied on Primary node during Physical Restore

When applying oplog chunks on the the Primary node, it might happen that
Primary node is still not reached Primary state, simply because after its
restart, it is in transition towards Primary:
STARTUP -> STARTUP2 -> RECOVERING -> SECONDARY -> PRIMARY

To fix such a situation, PBM checks if the Primary state is reached. After
the Primay state is reached, oplog is applied on that Primary.