Pull up storage creation out of waitRestore func
When doing a physical restore mongodb is down, and that can cause side
effects when creating Storage type. Storage requires configuration, which
is stored in db.
The bug is fixed by refactoring to DI pattern: creating storage
dependency on a single place and injecting it in functions that
require it.